Rework higher-rank kinds post to jive with simplified subsumption
Simplified subsumption means that regeneralization will no longer
swizzle type variables past function arrows, which breaks the `ex2`
example in "The surprising rigidness of higher-rank kinds". This
patch repairs the breakage by tweaking `ex2` to only rearrange type
variables within a single `forall`. This has the added bonus of
making the presentation slightly simpler, so I'd argue that this
version of the post is actually easier to understand.

Fixes #7.